+++ frontend/src/app/(authenticated)/catalog/[participantId]/[endpointUrl]/[assetId]/[tab]/page.tsx
@@ -6,6 +6,7 @@
  * SPDX-License-Identifier: Elastic-2.0
  */
 'use client';
+import {use} from 'react';
 
 import {CenteredLoadingSpinner} from '@/components/loading-spinner';
 import {api} from '@/lib/api/client';
@@ -19,16 +20,15 @@ import {decodeParams} from '@/lib/utils/http-utils';
 import CatalogDataOfferDetailPageContent from './catalog-data-offer-detail-page';
 import PageContainer from '@/components/page-container';
 
-export default function CatalogDataOfferDetailPage({
-  params,
-}: {
-  params: {
+export default function CatalogDataOfferDetailPage(props: {
+  params: Promise<{
     participantId: string;
     endpointUrl: string;
     assetId: string;
     tab: string;
-  };
+  }>;
 }) {
+  const params = use(props.params);
   const breadcrumbItems = useBreadcrumbItems();
   const {participantId, endpointUrl, assetId} = decodeParams(params);